TB-step_2 create data class, create and print dictionary with words and counter
