+++ src/main/kotlin/console_version/LearnEnglishWords_TB_consoleVersion.kt
@@ -1,0 +1,31 @@
+package console_version
+
+import java.io.File
+
+data class Word(
+    val englishWord: String,
+    val translation: String,
+    val correctAnswerCount: Int = 0,
+)
+
+fun main() {
+
+    val wordsFile = File("words.txt")
+    wordsFile.createNewFile()
+
+    val dictionary: MutableList<Word> = mutableListOf()
+
+    val listOfWords = wordsFile.readLines()
+
+    for (line in listOfWords) {
+        val line = line.split("|")
+        val word = Word(
+            englishWord = line[0],
+            translation = line[1],
+            correctAnswerCount = line[2].toIntOrNull() ?: 0,
+        )
+        dictionary.add(word)
+    }
+
+    println(dictionary)
+}
