--- src/main/kotlin/console_version/step_1/Step_1.kt
@@ -1,15 +1,0 @@
-package console_version.step_1
-
-import java.io.File
-
-fun main() {
-
-    val wordsFile = File("words.txt")
-    wordsFile.createNewFile()
-
-    val listOfWords = wordsFile.readLines()
-
-    for (line in listOfWords) {
-        println(line)
-    }
-}
