+++ words.txt
@@ -1,0 +1,3 @@
+hello|привет|2
+dog|собака|3
+cat|кошка|
